+++ src/ru/majestic/thetown/view/counters/ICountWithMaxValueView.java
@@ -2,6 +2,6 @@ package ru.majestic.thetown.view.counters;
 
 public interface ICountWithMaxValueView extends ICountView {
    
-   public void onMaxValueChanged(int maxValue);
+   public void onMaxValueChanged(long maxValue);
    
 }
